[FIX] web_editor: update the media search bar layout
Before, depending on the screen size, it was not always clear for the
user that he can use the search bar to find library images.

The upload buttons and search bars have been switched in order to make
the search the first choice for the user. Also, make the arrow pointing
at the searchbar more understandable.

task-2446852

closes odoo/odoo#68821

X-original-commit: 3a23a1c2c8869439763eb864a65a9f391e27d1f6
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>