[IMP] website_crm : catch logged partner info on contact form
1) Improves user experience by filling contact form with logged partner info.
This facilitates the flow when logged.

The contact us form is filled with data from the partner linked
with logged in web visitor, logged_partner. Its data is used in priority.
Empty fields (e.g. no partner is logged) will be set to default values
in request.params if any. This way, previous behaviour is preserved (as well
as widget correct display behaviour).

2) This commit also adds a star next to the required field Subject in the contact
us form like for the rest of required fields of the form : s_website_form_mark
This improves coherence and clarity when filling the form.

Task-ID : 2241639
PR : #61126